Add further experimental extensions
As it turns out, my mental model for this library was entirely incorrect
-- observers in Incremental land always have a value (although not
directly accessible). This percludes usage of operators like `filter`,
etc., as it's an error not to immediately receive a value when calling
`observe`.

In that sense this library is a more limited version of a more
traditional reactive library -- but hopefully one that is still useful.